Refactored set-selection-from-dom into utils as prep for Android support
